Migrated dataframe_visualizer.py to as_dataframe.py
